Readded BLS
By adding other considerations (increased security) for the sharded nodes, BLS becomes an option again.
This allows to utilise BLS to re-shrink transactions to an all-time low.

Additionally the nonce is taken away from the transaction data. This saves another 4 byte (28->24) but also adds the risk of replay attacks.
This is fixed by hashing the current block height aswell as the transaction data when signing it. Since a transaction is included in the block where its submitted anyway, there is no need to introduce another variable.

And again, the usual:
Do not reuse your keys on other networks or forks!
Doing so will result in your creating transactions that can be replayed on the main network.